move reflector controller
