+++ pkg/reflectorcontroller/controller.go
@@ -1,4 +1,4 @@
-package reflector
+package reflectorcontroller
 
 import (
 	"sync"
@@ -6,23 +6,23 @@ import (
 	"k8s.io/client-go/tools/cache"
 )
 
-// Controller controls start and stop of multiple reflectors
-type Controller struct {
+// ReflectorController controls start and stop of multiple reflectors
+type ReflectorController struct {
 	reflectors []*cache.Reflector
 	mu         sync.Mutex
 	running    bool
 	stopCh     chan struct{}
 }
 
-// NewController creates a new Controller responsible for starting and stopping the reflectors specified in the input parameter
-func NewController(reflectors ...*cache.Reflector) *Controller {
-	return &Controller{
+// New creates a new ReflectorController responsible for starting and stopping the reflectors specified in the input parameter
+func New(reflectors ...*cache.Reflector) *ReflectorController {
+	return &ReflectorController{
 		reflectors: reflectors,
 	}
 }
 
 // Start all reflectors referenced by this controller, or no-op if already started.
-func (c *Controller) Start() {
+func (c *ReflectorController) Start() {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	if c.running {
@@ -35,7 +35,7 @@ func (c *Controller) Start() {
 }
 
 //Stop all reflectors referenced by this controller, or no-op if not running.
-func (c *Controller) Stop() {
+func (c *ReflectorController) Stop() {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	if !c.running {
